+++ src/collective/taxonomy/interfaces.py
@@ -67,12 +67,6 @@ def taxonomyDefaultValue():
     return taxonomy
 
 
-def get_lang_code(lang=None):
-    if lang is None:
-        lang = api.portal.get_current_language()
-    return lang.split('-', 1)[0]
-
-
 class ITaxonomyForm(Interface):
 
     taxonomy = schema.TextLine(
@@ -96,9 +90,9 @@ class ITaxonomyForm(Interface):
 
     default_language = schema.Choice(
         title=_(u"Default language"),
-        vocabulary='collective.taxonomy.languages',
+        vocabulary='plone.app.vocabularies.AvailableContentLanguages',
         required=True,
-        defaultFactory=get_lang_code
+        defaultFactory=api.portal.get_current_language
     )
 
     import_file = NamedBlobFile(